Allow /data to be remounted

+++ conf/provisioning/data_volume.sh
@@ -7,11 +7,17 @@ set -v
 
 mkdir /data
 
-lsblk
-parted /dev/sdb --script -- mklabel gpt
-parted /dev/sdb --script -- mkpart primary 0% 100%
-lsblk
-mkfs.ext4 /dev/sdb1
+if lsblk | fgrep sdb1; then
+ echo "sdb1 already exists, skipping particioning"
+else
+  lsblk
+  parted /dev/sdb --script -- mklabel gpt
+  parted /dev/sdb --script -- mkpart primary 0% 100%
+  lsblk
+  mkfs.ext4 /dev/sdb1
+fi
+
+
 echo "/dev/sdb1    /data     ext4    defaults    0    0" >> /etc/fstab
 
 mount /data
